+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "account-backend",
-  "version": "0.0.0",
+  "version": "0.2.0",
   "private": true,
   "type": "module",
   "license": "UNLICENSED",
@@ -71,6 +71,7 @@
     "pg": "^8.16.3",
     "proxy-addr": "^2.0.7",
     "reflect-metadata": "^0.2.2",
+    "smol-toml": "^1.6.0",
     "uuid": "^13.0.0"
   },
   "hotHook": {
