+++ start/kernel.ts
@@ -10,32 +10,13 @@
 
 import router from '@adonisjs/core/services/router'
 import server from '@adonisjs/core/services/server'
-import { readFile } from 'fs'
 
 /**
  * The error handler is used to convert an exception
  * to an HTTP response.
  */
 server.errorHandler(() => import('#exceptions/handler'))
-interface PdnodeConfig{
-  banned: {
-    username: Array<string>,
-    nickname: Array<string>
-  }
-}
-export let config: PdnodeConfig
 
-
-readFile("pdnode.config.json", "utf8", (err, data) => {
-  if(err){
-    console.log("Read Config File Failed: " + err)
-  }
-  try {
-      config = JSON.parse(data);
-    } catch (parseError) {
-      throw  console.error("Paese Config File Failed, Error: ", parseError);
-    }
-})
 
 /**
  * The server middleware stack runs middleware on all the HTTP
